serviceuser commit

+++ service/service_hosp/src/main/java/com/henu/hosp/controller/api/ApiController.java
@@ -63,7 +63,9 @@ public class ApiController {
 
         //医院编号
         String hoscode = (String)paramMap.get("hoscode");
-
+        if(StringUtils.isEmpty(hoscode)) {
+            throw new SumException(ResultCodeEnum.PARAM_ERROR);
+        }
         //科室编号
         String depcode = (String)paramMap.get("depcode");
         //当前页 和 每页记录数
@@ -85,8 +87,6 @@ public class ApiController {
         Map<String, String[]> requestMap = request.getParameterMap();
         Map<String, Object> paramMap = HttpRequestHelper.switchMap(requestMap);
         //将String数组转换为object类型
-
-
         scheduleService.save(paramMap);
         return Result.ok();
     }
@@ -102,7 +102,9 @@ public class ApiController {
         //医院编号 和 科室编号
         String hoscode = (String)paramMap.get("hoscode");
         String depcode = (String)paramMap.get("depcode");
-
+        if(StringUtils.isEmpty(hoscode)) {
+            throw new SumException(ResultCodeEnum.PARAM_ERROR);
+        }
         departmentService.remove(hoscode,depcode);
         return Result.ok();
     }
@@ -117,6 +119,9 @@ public class ApiController {
 
         //医院编号
         String hoscode = (String)paramMap.get("hoscode");
+        if(StringUtils.isEmpty(hoscode)) {
+            throw new SumException(ResultCodeEnum.PARAM_ERROR);
+        }
         //当前页 和 每页记录数
         int page = StringUtils.isEmpty(paramMap.get("page")) ? 1 : Integer.parseInt((String)paramMap.get("page"));
         int limit = StringUtils.isEmpty(paramMap.get("limit")) ? 1 : Integer.parseInt((String)paramMap.get("limit"));
@@ -140,7 +145,9 @@ public class ApiController {
         String hoscode = (String)paramMap.get("hoscode");
         //1 获取医院系统传递过来的签名,签名进行MD5加密
         String hospSign = (String)paramMap.get("sign");
-
+        if(StringUtils.isEmpty(hoscode)) {
+            throw new SumException(ResultCodeEnum.PARAM_ERROR);
+        }
         //2 根据传递过来医院编码，查询数据库，查询签名
         String signKey = hospitalSetService.getSignKey(hoscode);
 
@@ -199,11 +206,15 @@ public class ApiController {
 
         //2 根据传递过来医院编码，查询数据库，查询签名
         String hoscode = (String)paramMap.get("hoscode");
+
+        if(StringUtils.isEmpty(hoscode)) {
+            throw new SumException(ResultCodeEnum.PARAM_ERROR);
+        }
+
         String signKey = hospitalSetService.getSignKey(hoscode);
 
         //3 把数据库查询签名进行MD5加密
         String signKeyMd5 = MD5.encrypt(signKey);
-
         //4 判断签名是否一致
         if(!hospSign.equals(signKeyMd5)) {
             throw new SumException(ResultCodeEnum.SIGN_ERROR);
